requirements.txt
# Updated dependencies
python-telegram-bot==20.3
pymongo==4.12.0
psutil==7.0.0
flask==2.3.3
requests==2.31.0
dos2unix==1.0.0